+++ cli/completions/bash-completion.sh
@@ -170,6 +170,11 @@ _caupain() {
           in_param=''
           continue
           ;;
+        --do-not-check-self-updates)
+          __skip_opt_eq
+          in_param=''
+          continue
+          ;;
         --version)
           __skip_opt_eq
           in_param=''
@@ -192,7 +197,7 @@ _caupain() {
   done
   local word="${COMP_WORDS[$COMP_CWORD]}"
   if [[ "${word}" =~ ^[-] ]]; then
-    COMPREPLY=($(compgen -W '-i --version-catalog --gradle-wrapper-properties -e --excluded -c --config --policy-plugin-dir -p --policy --list-policies --gradle-stability-level -t --output-type -o --output --output-dir --output-base-name --show-version-references --in-place --github-token --search-release-notes --cache-dir --no-cache --clean-cache -q --quiet -v --verbose -d --debug --debug-http-calls --verify-existence --version -h --help' -- "${word}"))
+    COMPREPLY=($(compgen -W '-i --version-catalog --gradle-wrapper-properties -e --excluded -c --config --policy-plugin-dir -p --policy --list-policies --gradle-stability-level -t --output-type -o --output --output-dir --output-base-name --show-version-references --in-place --github-token --search-release-notes --cache-dir --no-cache --clean-cache -q --quiet -v --verbose -d --debug --debug-http-calls --verify-existence --do-not-check-self-updates --version -h --help' -- "${word}"))
     return
   fi
 
@@ -258,6 +263,8 @@ _caupain() {
     "--debug-http-calls")
       ;;
     "--verify-existence")
+      ;;
+    "--do-not-check-self-updates")
       ;;
     "--version")
       ;;